Moved ServiceGetter.java to org.apache.sling.junit.impl package
