+++ src/main/java/org/apache/sling/junit/rules/ServerSideTeleporter.java
@@ -20,7 +20,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.sling.junit.Activator;
-import org.apache.sling.junit.ServiceGetter;
+import org.apache.sling.junit.impl.ServiceGetter;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.FrameworkUtil;
