+++ src/main/java/org/apache/sling/junit/rules/Service.java
@@ -18,7 +18,7 @@
 package org.apache.sling.junit.rules;
 
 import org.apache.sling.junit.Activator;
-import org.apache.sling.junit.ServiceGetter;
+import org.apache.sling.junit.impl.ServiceGetter;
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
